# docs/openapi.json
{
  "openapi": "3.1.0",
  "info": {
    "title": "QuantSynthica Market API",
    "description": "Unified Financial-Data API integrating yfinance, TradingView-Screener, and Screener.in with quantitative analytics.",
    "version": "1.0.0"
  },
  "paths": {
...
    "/api/v1/valuation/{symbol}": {
      "get": {
        "tags": [
          "Valuation Models"
        ],
        "summary": "Get Valuation Overview",
        "description": "Return comprehensive valuation overview combining DCF, Piotroski F-Score, Altman Z-Score, and Graham Number.",
        "operationId": "get_valuation_overview_api_v1_valuation__symbol__get",
        "security": [
          {
            "APIKeyHeader": []
          }
        ],
        "parameters": [
          {
            "name": "symbol",
            "in": "path",
            "required": true,
            "schema": {
              "type": "string",
              "title": "Symbol"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "Successful Response",
            "content": {
              "application/json": {
                "schema": {
                  "type": "object",
                  "additionalProperties": true,
                  "title": "Response Get Valuation Overview Api V1 Valuation  Symbol  Get"
                }
              }
            }
          },
          "422": {
            "description": "Validation Error",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            }
          }
        }
      }
    },
    "/api/v1/valuation/dcf": {
      "post": {
        "tags": [
          "Valuation Models"
        ],
        "summary": "Run Dcf Valuation",
        "description": "Execute custom Discounted Cash Flow (DCF) model with custom growth rates, discount rate and terminal assumptions.",
        "operationId": "run_dcf_valuation_api_v1_valuation_dcf_post",
        "requestBody": {
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/DCFRequest"
              }
            }
          },
          "required": true
        },
        "responses": {
          "200": {
            "description": "Successful Response",
            "content": {
              "application/json": {
                "schema": {
                  "additionalProperties": true,
                  "type": "object",
                  "title": "Response Run Dcf Valuation Api V1 Valuation Dcf Post"
                }
              }
            }
          },
          "422": {
            "description": "Validation Error",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            }
          }
        },
        "security": [
          {
            "APIKeyHeader": []
          }
        ]
      }
    },
    "/api/v1/valuation/piotroski/{symbol}": {
      "get": {
        "tags": [
          "Valuation Models"
        ],
        "summary": "Get Piotroski F Score",
        "description": "Calculate the 9-point Piotroski F-Score for financial strength and accounting health.",
        "operationId": "get_piotroski_f_score_api_v1_valuation_piotroski__symbol__get",
        "security": [
          {
            "APIKeyHeader": []
          }
        ],
        "parameters": [
          {
            "name": "symbol",
            "in": "path",
            "required": true,
            "schema": {
              "type": "string",
              "title": "Symbol"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "Successful Response",
            "content": {
              "application/json": {
                "schema": {
                  "type": "object",
                  "additionalProperties": true,
                  "title": "Response Get Piotroski F Score Api V1 Valuation Piotroski  Symbol  Get"
                }
              }
            }
          },
          "422": {
            "description": "Validation Error",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            }
          }
        }
      }
    },
    "/api/v1/valuation/altman-z/{symbol}": {
      "get": {
        "tags": [
          "Valuation Models"
        ],
        "summary": "Get Altman Z Score",
        "description": "Calculate the Altman Z-Score for bankruptcy distress prediction and credit health.",
        "operationId": "get_altman_z_score_api_v1_valuation_altman_z__symbol__get",
        "security": [
          {
            "APIKeyHeader": []
          }
        ],
        "parameters": [
          {
            "name": "symbol",
            "in": "path",
            "required": true,
            "schema": {
              "type": "string",
              "title": "Symbol"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "Successful Response",
            "content": {
              "application/json": {
                "schema": {
                  "type": "object",
                  "additionalProperties": true,
                  "title": "Response Get Altman Z Score Api V1 Valuation Altman Z  Symbol  Get"
                }
              }
            }
          },
          "422": {
            "description": "Validation Error",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            }
          }
        }
      }
    },
    "/api/v1/valuation/graham/{symbol}": {
      "get": {
        "tags": [
          "Valuation Models"
        ],
        "summary": "Get Graham Valuation",
        "description": "Calculate Benjamin Graham Number and Net-Net Working Capital valuation.",
        "operationId": "get_graham_valuation_api_v1_valuation_graham__symbol__get",
        "security": [
          {
            "APIKeyHeader": []
          }
        ],
        "parameters": [
          {
            "name": "symbol",
            "in": "path",
            "required": true,
            "schema": {
              "type": "string",
              "title": "Symbol"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "Successful Response",
            "content": {
              "application/json": {
                "schema": {
                  "type": "object",
                  "additionalProperties": true,
                  "title": "Response Get Graham Valuation Api V1 Valuation Graham  Symbol  Get"
                }
              }
            }
          },
          "422": {
            "description": "Validation Error",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            }
          }
        }
      }
    },
...
    "/api/v1/sentiment/{symbol}": {
      "get": {
        "tags": [
          "Financial Sentiment & NLP"
        ],
        "summary": "Get Symbol Sentiment",
        "description": "Retrieve aggregate sentiment and article-level sentiment breakdown for recent company news.",
        "operationId": "get_symbol_sentiment_api_v1_sentiment__symbol__get",
        "security": [
          {
            "APIKeyHeader": []
          }
        ],
        "parameters": [
          {
            "name": "symbol",
            "in": "path",
            "required": true,
            "schema": {
              "type": "string",
              "title": "Symbol"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "Successful Response",
            "content": {
              "application/json": {
                "schema": {
                  "type": "object",
                  "additionalProperties": true,
                  "title": "Response Get Symbol Sentiment Api V1 Sentiment  Symbol  Get"
                }
              }
            }
          },
          "422": {
            "description": "Validation Error",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            }
          }
        }
      }
    },
    "/api/v1/sentiment/analyze": {
      "post": {
        "tags": [
          "Financial Sentiment & NLP"
        ],
        "summary": "Analyze Text Sentiment",
        "description": "Analyze sentiment polarity, compound score, and financial keywords for arbitrary text.",
        "operationId": "analyze_text_sentiment_api_v1_sentiment_analyze_post",
        "requestBody": {
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/AnalyzeTextRequest"
              }
            }
          },
          "required": true
        },
        "responses": {
          "200": {
            "description": "Successful Response",
            "content": {
              "application/json": {
                "schema": {
                  "additionalProperties": true,
                  "type": "object",
                  "title": "Response Analyze Text Sentiment Api V1 Sentiment Analyze Post"
                }
              }
            }
          },
          "422": {
            "description": "Validation Error",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            }
          }
        },
        "security": [
          {
            "APIKeyHeader": []
          }
        ]
      }
    },
...
  },
  "components": {
    "schemas": {
...
      "AnalyzeTextRequest": {
        "properties": {
          "text": {
            "type": "string",
            "title": "Text",
            "description": "Financial text, article snippet, earnings commentary, or headline to analyze"
          }
        },
        "type": "object",
        "required": [
          "text"
        ],
        "title": "AnalyzeTextRequest"
      },
...
      "DCFRequest": {
        "properties": {
          "symbol": {
            "type": "string",
            "title": "Symbol",
            "description": "Stock symbol (e.g. RELIANCE, AAPL)"
          },
          "growth_rate": {
            "anyOf": [
              {
                "type": "number"
              },
              {
                "type": "null"
              }
            ],
            "title": "Growth Rate",
            "description": "Projected annual FCF growth rate (e.g. 0.12 for 12%)"
          },
          "discount_rate": {
            "anyOf": [
              {
                "type": "number"
              },
              {
                "type": "null"
              }
            ],
            "title": "Discount Rate",
            "description": "Discount rate / WACC (e.g. 0.10 for 10%)"
          },
          "terminal_growth_rate": {
            "anyOf": [
              {
                "type": "number"
              },
              {
                "type": "null"
              }
            ],
            "title": "Terminal Growth Rate",
            "description": "Terminal perpetual growth rate (e.g. 0.03 for 3%)"
          },
          "years": {
            "type": "integer",
            "title": "Years",
            "description": "Projection years (default 5)",
            "default": 5
          },
          "exit_multiple": {
            "anyOf": [
              {
                "type": "number"
              },
              {
                "type": "null"
              }
            ],
            "title": "Exit Multiple",
            "description": "Optional EV/EBITDA or EV/FCF exit multiple"
          }
        },
        "type": "object",
        "required": [
          "symbol"
        ],
        "title": "DCFRequest"
      },
...
      "HTTPValidationError": {
        "properties": {
          "detail": {
            "items": {
              "$ref": "#/components/schemas/ValidationError"
            },
            "type": "array",
            "title": "Detail"
          }
        },
        "type": "object",
        "title": "HTTPValidationError"
      },
...
      "ValidationError": {
        "properties": {
          "loc": {
            "items": {
              "anyOf": [
                {
                  "type": "string"
                },
                {
                  "type": "integer"
                }
              ]
            },
            "type": "array",
            "title": "Location"
          },
          "msg": {
            "type": "string",
            "title": "Message"
          },
          "type": {
            "type": "string",
            "title": "Error Type"
          },
          "input": {
            "title": "Input"
          },
          "ctx": {
            "type": "object",
            "title": "Context"
          }
        },
        "type": "object",
        "required": [
          "loc",
          "msg",
          "type"
        ],
        "title": "ValidationError"
      },
...
    },
    "securitySchemes": {
      "APIKeyHeader": {
        "type": "apiKey",
        "in": "header",
        "name": "X-API-Key"
      }
    }
  }
}